backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
